+++ source/PhysicalQuantity/Mass.php
@@ -1,11 +1,11 @@
 <?php
 namespace PhpUnitsOfMeasure\PhysicalQuantity;
 
-use PhpUnitsOfMeasure\PhysicalQuantity;
+use PhpUnitsOfMeasure\BasePhysicalQuantity;
 use PhpUnitsOfMeasure\UnitOfMeasure;
 use PhpUnitsOfMeasure\HasSIUnitsTrait;
 
-class Mass extends PhysicalQuantity
+class Mass extends BasePhysicalQuantity
 {
     use HasSIUnitsTrait;
 
@@ -13,13 +13,15 @@ class Mass extends PhysicalQuantity
 
     static protected $hasBeenInitialized = false;
 
+    static protected $nativeUnitOfMeasure;
+
     static protected function registerDefaultUnitsOfMeasure()
     {
         // Kilogram
         $kilogram = UnitOfMeasure::nativeUnitFactory('kg');
         $kilogram->addAlias('kilogram');
         $kilogram->addAlias('kilograms');
-        static::registerUnitOfMeasure($kilogram);
+        static::registerNativeUnitOfMeasure($kilogram);
 
         static::addMissingSIPrefixedUnits(
             $kilogram,