+++ tests/PhysicalQuantity/DimensionliessCoefficientTest.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace PhpUnitsOfMeasureTest\PhysicalQuantity;
+
+use PhpUnitsOfMeasure\PhysicalQuantity\DimensionlessCoefficient;
+
+class DimensionlessCoefficientTest extends AbstractPhysicalQuantityTestCase
+{
+    protected $supportedUnitsWithAliases = [
+        '',
+    ];
+
+    protected function instantiateTestQuantity()
+    {
+        return new DimensionlessCoefficient(1);
+    }
+}